🚚 Moved Module and Library types out of stdlib into its own file
